third_party/blink/web_tests/virtual/prefetch/external/wpt/speculation-rules/prefetch/tentative/service-worker/redirect.sub.https_origin=same-site_sc=in-in_sw=fetch-handler-to-fallback-expected.txt

This is a testharness.js-based test.
[FAIL] Prefetch with ServiceWorker (fetch-handler-to-fallback)
  assert_equals: prefetch requests to initial URL should be sent to the server. expected 1 but got 0
Harness: the test ran to completion.
